Move implementations to internal/
